pyre-jit-trace: extract branch + resume-snapshot helpers
Move the branch-target resolution / trampoline decoding / branch-arm
resume-liveness helpers into branch.rs, and the guard resume-data
snapshot capture (single- and multi-frame inline paths plus the
inline-caller-frame computation) into resume_snapshot.rs.
